ml-logic-service/src/requirements.txt
annotated-types==0.7.0
async-timeout==5.0.1
cachetools==5.5.0
camel-case-switcher==2.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.5.14
filelock==3.13.1
fsspec==2024.2.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.67.1
grpcio-status==1.67.1
http-server-base==2.0.10
httplib2==0.22.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.5
marshmallow==3.23.1
mpmath==1.3.0
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
networkx==3.2.1
numpy==1.26.3
packaging==24.2
parse==1.19.1
pillow==10.2.0
proto-plus==1.25.0
protobuf==4.21.12
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
supervisor==4.2.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.2
typing-inspect==0.7.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.3
